scripts/storage/requirements.txt
grpcio==1.15.0
grpcio-tools==1.15.0
protobuf==3.6.1
tqdm==4.19.5
